backend/requirements.txt
# FastAPI y dependencias principales
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic[email]==2.5.0
pydantic-settings==2.1.0

# Base de datos
motor==3.3.2
pymongo==4.6.0
redis==5.0.1

# Autenticación y seguridad
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
PyJWT==2.8.0

# Manejo de archivos y multimedia
boto3==1.34.0
Pillow==10.1.0
python-magic==0.4.27

# Geocodificación y mapas
geopy==2.4.1
shapely==2.0.2

# Utilidades
python-dotenv==1.0.0
httpx==0.25.2
celery==5.3.4

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Logging y monitoreo
structlog==23.2.0
